spark/workspace/requirements.txt
requests
confluent-kafka
delta-spark==1.0.1